draw frame based off of time value
Changes draw_frame to take in a time and then derives text position from
the time value. I was able to work out most of the math and
troublehsooting with the simulator and then flashing was pretty easy
except that I forgot to clear the display since that happens outside the
draw_frame funciton until I work out the trait requirements.

Everything looks good in the simulator but running on the
microcontroller + display is glacial. I think this is because of the
modulo math in the draw funciton. I might have to go back to passing in
a Scene and drawing based on that. Cool example of real world
performance constraints.